Introduce and integrate load command visitor
This infra will be used as the first pass to ingest data from input
files. The infra isn't ready yet for reasons listed below, but for now
it correctly parses load commands and calls the right vtable entry.

Below is the TODO list for this visitor:

- Pass through the file name when visiting a mach_header
  - Easiest to introduce the a new member "path" to ald::MachO::File
- Somehow create context so that a visitor knows that a given load
    command belongs to a given mach header that was previously parsed
    (the base visitor should hold whatever state is necessary to achieve
     this). Special care should be taken to ensure this is thread safe.
- Create a MachO file visitor that does common parsing tasks like:
  - Parse each section within a segment command
  - Extract MemoryBuffer objects for each section/segment

With the infra proposed above we can move onto implementing the initial
consume/map pass. Some design decisions will need to be made about when
to "reduce" i.e. when to combine the information that each file
contains.