Pass a::M::File through LCVisitor
Now an LCVisitor knows which file is getting parsed when its callbacks
are being called. The TODOs now are to connect each load command
callback to a given file and do common parsing tasks (parse
segments, sections, create MemoryBuffer for each). The latter depends on
the former so that we can incrementally introduce new features/parsing
via composition of visitors (i.e. we teach each visitor which File is
being parsed when a given callback is called and then we can
implement a visitor that overrides visiting a segment to parse the
sections and generate memory buffers- note the MB require the File context so
that the visitor can determine where in the memory the data lies).